applications/proteins/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=0.15.0=pypi_0
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.4.0=pypi_0
aom=3.6.0=h6a678d5_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
binutils_impl_linux-64=2.40=ha1999f0_7
biopython=1.84=pypi_0
biotite=0.36.1=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
boost-cpp=1.69.0=h11c811c_1000
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2024.7.4=py39h06a4308_0
cffi=1.17.0=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
comm=0.2.2=pypi_0
configparser=7.1.0=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.2.1=pypi_0
cssselect=1.2.0=pypi_0
cssutils=2.11.1=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.12.1=pypi_0
dataframe-image=0.2.4=pypi_0
dav1d=1.2.1=h5eee18b_0
debugpy=1.8.5=pypi_0
decorator=5.1.0=pypi_0
deepspeed=0.15.0=pypi_0
defusedxml=0.7.1=pypi_0
dgl-cuda11.1=0.9.1post1=py39_0
dllogger=1.0.0=pypi_0
dm-tree=0.1.7=pypi_0
dssp=3.0.0=h3fd9d12_4
e3nn=0.3.3=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.0.1=pypi_0
expat=2.6.2=h6a678d5_0
fair-esm=2.0.0=pypi_0
fastjsonschema=2.20.0=pypi_0
ffmpeg=6.1.1=h4c62175_0
filelock=3.15.4=pypi_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.53.1=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
gcc=14.1.0=h6f9ffa1_0
gcc_impl_linux-64=14.1.0=h3c94d91_0
generate-chroma=1.0.0=pypi_0
giflib=5.2.1=h5eee18b_3
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gputil=1.4.0=pypi_0
graphite2=1.3.14=h295c915_1
h11=0.14.0=pypi_0
harfbuzz=4.3.0=hf52aaf7_1
hjson=3.1.0=pypi_0
html2image=2.0.4.3=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.24.6=pypi_0
hydra-core=1.3.2=pypi_0
hydra-joblib-launcher=1.2.0=pypi_0
icu=58.2=he6710b0_3
idna=3.7=py39h06a4308_0
importlib-metadata=8.4.0=pypi_0
importlib-resources=6.4.4=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.5=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=py39h06a4308_0
joblib=1.4.2=py39h06a4308_0
jpeg=9e=h5eee18b_3
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.5=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kernel-headers_linux-64=2.6.32=he073ed8_17
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=hf3520f5_7
leptonica=1.82.0=h42c8aad_2
lerc=3.0=h295c915_0
libarchive=3.6.2=h6ac8c49_3
libboost=1.73.0=h28710b8_12
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-devel_linux-64=14.1.0=h5d3d1c9_100
libgcc-ng=14.1.0=h77fa898_0
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=14.1.0=h77fa898_0
libiconv=1.16=h5eee18b_3
libogg=1.3.5=h27cfd23_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libsanitizer=14.1.0=hcba0ae0_0
libstdcxx-ng=14.1.0=hc0a3c3a_0
libtheora=1.1.1=h7f8727e_3
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libuv=1.48.0=h5eee18b_0
libvorbis=1.3.7=h7b6447c_0
libvpx=1.13.1=h6a678d5_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxml2=2.10.4=hcbfbd50_0
lxml=5.3.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdtraj=1.10.0=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
ml-collections=0.1.1=pypi_0
more-itertools=10.4.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
multidict=6.0.5=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.2.1=py39h06a4308_0
nglview=3.1.2=pypi_0
ninja=1.11.1.1=pypi_0
notebook=7.2.2=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
ocl-icd=2.3.2=h5eee18b_1
ocl-icd-system=1.0.0=1
omegaconf=2.3.0=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openmm=8.0.0=py39h31b9435_1
openssl=3.0.14=h5eee18b_0
opt-einsum=3.3.0=pypi_0
opt-einsum-fx=0.1.4=pypi_0
overrides=7.7.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pcre2=10.42=hebb0a14_1
pdbfixer=1.9.0=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=py39h5eee18b_0
pip=24.2=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
plotly=5.23.0=pypi_0
pluggy=1.5.0=pypi_0
prometheus-client=0.20.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.47=pypi_0
proteus=0.0.0=dev_0
protobuf=5.27.3=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
pybind11-abi=4=hd3eb1b0_1
pycparser=2.22=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pyg=2.1.0=py39_torch_1.11.0_cu102
pygments=2.18.0=pypi_0
pynvml=11.0.0=pypi_0
pyparsing=3.1.2=pypi_0
pyrsistent=0.20.0=pypi_0
pysocks=1.7.1=py39h06a4308_0
pytest=8.3.2=pypi_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda10.2_cudnn7.6.5_0
pytorch-cluster=1.6.0=py39_torch_1.11.0_cu102
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py39_torch_1.11.0_cu102
pytorch-sparse=0.6.15=py39_torch_1.11.0_cu102
pytz=2024.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.7.24=pypi_0
requests=2.32.3=py39h06a4308_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rfdiffusion=1.1.0=dev_0
rpds-py=0.20.0=pypi_0
safetensors=0.4.4=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.13.1=py39h5f9d8c6_0
se3-diffusion=0.0.0=dev_0
send2trash=1.8.3=pypi_0
setuptools=72.1.0=py39h06a4308_0
shortuuid=1.0.13=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
subprocess32=3.5.4=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.13.2=pypi_0
sysroot_linux-64=2.12=he073ed8_17
tbb=2021.8.0=hdb19cb5_0
tenacity=9.0.0=pypi_0
terminado=0.18.1=pypi_0
tesseract=5.2.0=h6a678d5_0
threadpoolctl=3.5.0=py39h2f386ee_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h39e8969_0
tmtools=0.0.2=pypi_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
torchaudio=0.11.0=py39_cu102
torchvision=0.12.0=py39_cu102
tqdm=4.66.4=py39h2f386ee_0
transformers=4.24.0=pypi_0
tree=0.2.4=pypi_0
types-python-dateutil=2.9.0.20240821=pypi_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.2=py39h06a4308_0
wandb=0.12.0=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.8.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.43.0=py39h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.9.4=pypi_0
zipp=3.20.1=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
